% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power}
\alias{power}
\alias{power_3d}
\title{power atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
power

power_3d
}
\description{
Brain-wide putative functional areas from resting state.
}
\examples{
data(power)
data(power_3d)
}
\references{
Power, J. D., Cohen, A. L., Nelson, S. M., Wig, G. S., Barnes, K. A., 
    Church, J. A., ... & Petersen, S. E. (2011). Functional network organization 
    of the human brain. Neuron, 72(4), 665-678.
    (\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3222858/}{PubMed})
    
\itemize{
 \item{power}{ - power atlas}
 \item{power_3d}{ - power 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
