% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{power}
\alias{power}
\title{Power Functional Network Cortical Parcellation}
\usage{
power()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Power cortical parcellation with 129 regions
based on functional network organization of the human brain.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
power()
}
\references{
Power JD, Cohen AL, Nelson SM, Wig GS, Barnes KA,
Church JA, Vogel AC, Laumann TO, Miezin FM, Schlaggar BL,
Petersen SE (2011). Functional Network Organization of the Human
Brain. \emph{Neuron}, 72(4):665-678.
\doi{10.1016/j.neuron.2011.09.006}
}
\concept{ggseg_atlases}
